Remove nesting
