+++ src/App.js
@@ -3,7 +3,6 @@ import { connect } from 'react-redux';
 import './App.css';
 import { BrowserRouter as Router} from 'react-router-dom';
 import Navbar from './components/Navbar';
-import Home from './components/Home';
 import * as actions from './state/actionCreators';
 
 function App(props) {
@@ -14,12 +13,6 @@ function App(props) {
     </div>
     </Router>
   );
-}
-
-const mapStateToProps = state => {
-  return {
-    state,
-  }
 }
 
 export default connect(state=>state, actions)(App);
